Fix warning from clang on testing function return (by Scott)
m_socket->send() returns a bool, but code in ServerSocket.cpp compares
it against -1.  clang16 generates a warning that this comparision
is never true.

Apparently this is written this way because the library used to return
an int here, but now signals errors by raising an exception.
Avoid the warning by casting to an int.